+++ packages/ui/modules/components/widgets/vanilla/core/VanillaButton.jsx
@@ -1,13 +1,14 @@
 import React from "react";
 
-export default ({type, label, onClick, readonly, config}) => {
-  const typeToLabel = {
-    "addRuleGroup": "+",
-    "addRuleGroupExt": "+",
-    "delGroup": "x",
-    "delRuleGroup": "x",
-    "delRule": "x",
-  };
+const typeToLabel = {
+  "addRuleGroup": "+",
+  "addRuleGroupExt": "+",
+  "delGroup": "x",
+  "delRuleGroup": "x",
+  "delRule": "x",
+};
+
+export default ({type, label, onClick, readonly}) => {
   const btnLabel = label || typeToLabel[type];
   return <button onClick={onClick} type="button" disabled={readonly}>{btnLabel}</button>;
 };