+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - Fixed import of rule_group in rule_group from SpEL (PR #959)
   - Updated type `ItemBuilderProps` (PR #959)
   - Fixed drag-n-drop to respect `maxNesting` when moving group into group (PR #959)
+  - Fix: allow custom conjuction in JsonLogic (issue #317) (PR #959)
 - 6.4.0
   - Functions can be used in LHS with `fieldSources: ["field", "func"]` in `settings`
     Thanks @rhallerman1 (PR #900, #896) (issues #287, #250, #344, #336)